requirements.txt
pandas
numpy
scikit-learn
lightgbm
shap
pyyaml
streamlit
plotly
joblib
